# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Loan CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /installments:
    get:
      operationId: listInstallments
      tags: [installments]
      summary: List installments with filters
      parameters:
        - name: client_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pendente, pago, atrasado]
        - name: vendedor_id
          in: query
          schema:
            type: integer
        - name: overdue_only
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of installments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Installment"
  /installments/{id}:
    get:
      operationId: getInstallment
      tags: [installments]
      summary: Get installment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Installment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installment"
        "404":
          description: Not found
    patch:
      operationId: updateInstallment
      tags: [installments]
      summary: Update installment (mark as paid, update value, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallmentUpdate"
      responses:
        "200":
          description: Updated installment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installment"
  /installments/{id}/pay:
    post:
      operationId: payInstallment
      tags: [installments]
      summary: Mark installment as paid and create event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: Payment registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installment"
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks for current user
      parameters:
        - name: tipo
          in: query
          schema:
            type: string
            enum: [followup, lembrete, pessoal]
        - name: concluido
          in: query
          schema:
            type: boolean
        - name: client_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Mark task as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  # PRIVATE NOTES
...
components:
  schemas:
...
    Installment:
      type: object
      required: [id, client_id, numero_parcela, valor, vencimento, status]
      properties:
        id:
          type: integer
        client_id:
          type: integer
        client_nome:
          type: ["string", "null"]
        vendedor_id:
          type: ["integer", "null"]
        numero_parcela:
          type: integer
        valor:
          type: number
        vencimento:
          type: string
        status:
          type: string
          enum: [pendente, pago, atrasado]
        pago_em:
          type: ["string", "null"]
    InstallmentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pendente, pago, atrasado]
        valor:
          type: number
        vencimento:
          type: string
    PaymentInput:
      type: object
      required: [valor_pago]
      properties:
        valor_pago:
          type: number
        observacao:
          type: string
...
    Task:
      type: object
      required: [id, user_id, titulo, data, tipo, concluido]
      properties:
        id:
          type: integer
        user_id:
          type: integer
        client_id:
          type: ["integer", "null"]
        client_nome:
          type: ["string", "null"]
        titulo:
          type: string
        data:
          type: string
        tipo:
          type: string
          enum: [followup, lembrete, pessoal]
        concluido:
          type: boolean
    TaskInput:
      type: object
      required: [titulo, data, tipo]
      properties:
        titulo:
          type: string
        data:
          type: string
        tipo:
          type: string
          enum: [followup, lembrete, pessoal]
        client_id:
          type: ["integer", "null"]
    TaskUpdate:
      type: object
      properties:
        titulo:
          type: string
        data:
          type: string
        tipo:
          type: string
          enum: [followup, lembrete, pessoal]
        concluido:
          type: boolean

    # PRIVATE NOTES
...